re-structure utils
